Throttle progress bar announcements
Limit how often the progress bar ARIA status can be updated to limit
how often the announcements happen.

If there hasn't been an announcement for a while, announce the update
immediately. If there was an announcement recently, schedule another
one. If an announcement is already scheduled, update the value so that
when the announcement happens it uses the latest progress value.